requirements.txt
python-telegram-bot>=20.0
aiohttp>=3.8.0

# Optional: Google Generative AI for generate_folders tool
google-generativeai>=0.7.2